backend/requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.2
FastAPI-SQLAlchemy==0.2.1
fastapi-users==9.3.1
fastapi-users-db-sqlalchemy==3.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
makefun==1.13.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
